moving test dataset to test utils
